+++ main/CMakeLists.txt
@@ -87,8 +87,8 @@ elseif(CONFIG_BOARD_TYPE_ATK_DNESP32S3_BOX)
     set(BOARD_TYPE "atk-dnesp32s3-box") 
 elseif(CONFIG_BOARD_TYPE_DU_CHATX)
     set(BOARD_TYPE "du-chatx")     
-elseif(CONFIG_BOARD_TYPE_PEILIAO_C3)
-    set(BOARD_TYPE "xiaozhipeiliao-c3")
+elseif(CONFIG_BOARD_TYPE_YUNLIAO_C3)
+    set(BOARD_TYPE "xiaozhiyunliao-c3")
 endif()
 file(GLOB BOARD_SOURCES ${CMAKE_CURRENT_SOURCE_DIR}/boards/${BOARD_TYPE}/*.cc)
 list(APPEND SOURCES ${BOARD_SOURCES})
