+++ main/boards/xiaozhiyunliao-c3/xiaoziyunliao_display.h
@@ -1,5 +1,5 @@
-#ifndef XIAOZIPEILIAO_DISPLAY_H
-#define XIAOZIPEILIAO_DISPLAY_H
+#ifndef XIAOZIYUNLIAO_DISPLAY_H
+#define XIAOZIYUNLIAO_DISPLAY_H
 
 #include "display/lcd_display.h"
 
@@ -14,14 +14,14 @@
 
 #include <atomic>
 
-class XiaozipeiliaoDisplay : public LcdDisplay {
+class XiaoziyunliaoDisplay : public LcdDisplay {
 public:
-    XiaozipeiliaoDisplay(esp_lcd_panel_io_handle_t panel_io, esp_lcd_panel_handle_t panel,
+    XiaoziyunliaoDisplay(esp_lcd_panel_io_handle_t panel_io, esp_lcd_panel_handle_t panel,
                   gpio_num_t backlight_pin, bool backlight_output_invert,
                   int width, int height,  int offset_x, int offset_y, bool mirror_x, bool mirror_y, bool swap_xy,
                   DisplayFonts fonts);
     void SetupUI() override;  // 添加方法重载声明
-    ~XiaozipeiliaoDisplay() override = default; // 显式声明析构函数
+    ~XiaoziyunliaoDisplay() override = default; // 显式声明析构函数
 
     // 添加虚函数覆盖声明
     void SetChatMessage(const std::string &role, const std::string &content) override;
@@ -32,4 +32,4 @@ protected:
 
 };
 
-#endif // XIAOZIPEILIAO_DISPLAY_H 
+#endif // XIAOZIYUNLIAO_DISPLAY_H
